Split dmx/engine.rs into submodules
- engine/playback.rs: play(), song time tracking, timeline wait
- engine/timeline.rs: timeline lifecycle, lighting validation, effect groups
- engine/midi_playback.rs: MIDI DMX playback cursor advancement
- Engine struct fields made pub(super) for submodule access
- Main engine.rs retains constructor, effects loop, MIDI event handling,
  universe management, public accessors, Drop impl, and all tests
